added 2fa app 2

+++ mfa/src/login.php
@@ -13,16 +13,18 @@ $pass = $_POST['password'] ?? '';
 if ($user === 'admin' && $pass === '1234') {
     $_SESSION['user'] = $user;
 
-    // Generar clave secreta única por sesión
-    $g = new GoogleAuthenticator();
-    $secret = $g->generateSecret();
-    $_SESSION['secret'] = $secret;
+    if(!isset($_SESSION['secret'])) { // Debe revisarse si el secreto no existe en la BD
+        // Generar clave secreta única por sesión
+        $g = new GoogleAuthenticator();
+        $secret = $g->generateSecret();
+        $_SESSION['secret'] = $secret;
 
-    // Mostrar QR
-    $qrCodeUrl = GoogleQrUrl::generate($user, $secret, 'MiSistemaMFA');
+        // Mostrar QR
+        $qrCodeUrl = GoogleQrUrl::generate($user, $secret, 'MiSistemaMFA');
 
-    echo "<h2>Escanea el código QR con Google Authenticator</h2>";
-    echo "<img src='$qrCodeUrl' />";
+        echo "<h2>Escanea el código QR con Google Authenticator</h2>";
+        echo "<img src='$qrCodeUrl' />";
+    }
     echo '<form action="verify.php" method="POST">
             <input type="text" name="code" placeholder="Código MFA" required>
             <button type="submit">Verificar</button>
